move timeseries functions from composition to their own library
