+++ .pylintrc
@@ -66,7 +66,6 @@ ignore=injected_glc_phosphorylation.py,
        agents_multigen.py,
        experiment.py,
        burst.py,
-       composition.py,
        dict_utils.py,
        control.py,
        emitter.py,
